End of module 12
They basically said  this is how it looks and works done with this module and the whole course. a lot of the code to do it was blocked out and some other the scripts needed to make things work were not there/ they did not how us where to get it.